+++ src/ncc/ThirdParty/Symfony/Process/PhpProcess.php
@@ -15,7 +15,7 @@ use ncc\ThirdParty\Symfony\Process\Exception\LogicException;
 use ncc\ThirdParty\Symfony\Process\Exception\RuntimeException;
 
 /**
- * PhpProcess runs a PHP script in an independent Process.
+ * PhpProcess runs a PHP script in an independent process.
  *
  *     $p = new PhpProcess('<?php echo "foo"; ?>');
  *     $p->run();
@@ -27,8 +27,8 @@ class PhpProcess extends Process
 {
     /**
      * @param string      $script  The PHP script to run (as a string)
-     * @param string|null $cwd     The working directory or null to use the working dir of the current PHP Process
-     * @param array|null  $env     The environment variables or null to use the same environment as the current PHP Process
+     * @param string|null $cwd     The working directory or null to use the working dir of the current PHP process
+     * @param array|null  $env     The environment variables or null to use the same environment as the current PHP process
      * @param int         $timeout The timeout in seconds
      * @param array|null  $php     Path to the PHP binary to use with any additional arguments
      */
@@ -50,17 +50,11 @@ class PhpProcess extends Process
         parent::__construct($php, $cwd, $env, $script, $timeout);
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public static function fromShellCommandline(string $command, string $cwd = null, array $env = null, mixed $input = null, ?float $timeout = 60): static
     {
         throw new LogicException(sprintf('The "%s()" method cannot be called when using "%s".', __METHOD__, self::class));
     }
 
-    /**
-     * {@inheritdoc}
-     */
     public function start(callable $callback = null, array $env = [])
     {
         if (null === $this->getCommandLine()) {