+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Updated `Symfony\polyfill-ctype` to version 1.27.0
  - Updated `Symfony\polyfill-mbstring` to version 1.27.0
  - Updated `Symfony\polyfill-uuid` to version 1.27.0
+ - Updated `Symfony\Process` to version 6.2.5
 
 ## [1.0.1] - 2023-02-07